REF: Move Formatter to printing
